+++ test/asset/outboxProject/srv/service/service.js
@@ -9,7 +9,7 @@ class NotificationService extends cds.Service {
       cds.log("sendFiori").info("sendFiori action triggered", {
         data: req.data,
         user: req.user.id,
-        subType: req.context._eventQueue?.queueEntries[0].subType,
+        subType: req.eventQueue?.processor.eventSubType,
       });
     });
 
@@ -18,7 +18,7 @@ class NotificationService extends cds.Service {
     });
 
     this.on("returnStatusAsArray", (req) => {
-      return [[req.context._eventQueue?.queueEntries[0].ID, req.data.status]];
+      return [[req.eventQueue.queueEntries[0].ID, req.data.status]];
     });
 
     this.on("rejectEvent", (req) => {